requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.10
django-br-utils==0.4
djangorestframework==3.16.1
pycparser==3.0
python-stdnum==2.2
sqlparse==0.5.5
